Update protobuf to v25.1 (#183)
# Changes

Since we are upgrading the protobuf version, the location for the root
CMakeLists.txt has changed, hence the update.

We also updated the swift targets (ex: `swift_add_library` or
`swift_add_executable`) so that we can now set a global default C++
version with `SWIFT_CXX_STANDARD`.

Lastly tried to unify the interface between `swift_add_test` and the
rest of the other swift targets since `swift_add_test` was developed
separately to the rest, hence the difference in interface.

These changes will allow us to set the C++ for our Swift targets
independently from the third party libraries.